chore(Dockerfile): update Splunk Forwarder config file path
Changed the path for copying configuration files to the Splunk Forwarder container to ensure they are placed in the correct directory (/opt/splunkforwarder/etc/system/local/) for proper configuration. This adjustment is meant to resolve an execution error during build time, whose remedial effects have yet to be tested, and may be reverted.